refactor settings
