+++ code/labflow-frontend/src/routes/instellingen/+page.svelte
@@ -6,12 +6,12 @@
     import GiTestTubes from 'svelte-icons/gi/GiTestTubes.svelte';
     // @ts-ignore
     import FaUserAlt from 'svelte-icons/fa/FaUserAlt.svelte';
-    import TestsBeheren from '../../components/Instellingen/TestsBeheren.svelte';
     import CategorieënBeheren from '../../components/Instellingen/CategorieënBeheren.svelte';
     import GebruikersBeheren from '../../components/Instellingen/GebruikersBeheren.svelte';
+	import StalenBeheren from '../../components/Instellingen/StalenBeheren.svelte';
 
     // default staat nu op testen beheren
-    let navElement = "Categorie";
+    let navElement = "Staal";
 </script>
 
 <Nav/>
@@ -23,9 +23,9 @@
             <div class="w-14 h-14 mb-5"><FaClipboardList /></div>
             Categorieën beheren
         </button>
-        <button type="button" on:click={() => navElement = "Test"} class="p-3 text-center text-lg bg-blue-400 flex flex-col items-center justify-center w-40 h-40 rounded-2xl text-white">
+        <button type="button" on:click={() => navElement = "Staal"} class="p-3 text-center text-lg bg-blue-400 flex flex-col items-center justify-center w-40 h-40 rounded-2xl text-white">
             <div class="w-14 h-14 mb-5"><GiTestTubes /></div>
-            Testen beheren
+            Stalen beheren
         </button>
         <button type="button" on:click={() => navElement = "Gebruiker"} class="p-3 text-center text-lg bg-blue-400 flex flex-col items-center justify-center w-40 h-40 rounded-2xl text-white">
             <div class="w-14 h-14 mb-5"><FaUserAlt /></div>
@@ -36,8 +36,8 @@
     <!-- content -->
     {#if navElement === "Categorie"}
         <CategorieënBeheren/>
-    {:else if navElement === "Test"}
-        <TestsBeheren/>
+    {:else if navElement === "Staal"}
+        <StalenBeheren/>
     {:else if navElement === "Gebruiker"}
         <GebruikersBeheren/>
     {/if}
